fix(webdav): 修复 payload_limit 类型和跨平台兼容性问题
- 将 payload_limit 类型从 usize 改为 u64，避免 32 位平台配置限制
- 添加 u64 到 usize 的安全转换逻辑，超出平台范围时使用 usize::MAX
- 在 Windows 交叉编译中添加 nasm 依赖以支持构建

+++ src/config/schema.rs
@@ -244,7 +244,7 @@ pub struct WebDavConfig {
     pub prefix: String,
     /// actix payload 硬上限，改了要重启。运行时软限制从 DB 读。
     #[serde(default = "WebDavConfig::default_payload_limit")]
-    pub payload_limit: usize,
+    pub payload_limit: u64,
 }
 
 impl Default for WebDavConfig {
@@ -260,7 +260,7 @@ impl WebDavConfig {
     fn default_prefix() -> String {
         "/webdav".to_string()
     }
-    fn default_payload_limit() -> usize {
+    fn default_payload_limit() -> u64 {
         10_737_418_240 // 10 GB 硬上限
     }
 }
